FileItemの不要な継承関係をなくすために変更
